[Ionut] - Replaced GLEW with glbinding (requires a CEGUI recompile. It currently crashes) - Replaced most "SceneGraphNode* const sgn" expressions with "SceneGraphNode& sgn" - Replaced some pointers to kernel to references to kernel - Fixed compilation issues with EASTL - Removed NPOT -> POT texture conversion code. Modern GPUs handle NPOT textures perfectly fine
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@354 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 50bee6dbab9c977c91187fd3c41d2c0f225a28a6 [formerly c946ad7d6054c767d11d93208c67d2695333d836] [formerly b3e9696e6dd5162065a237f85193b57948ae0b77 [formerly ce4678f307e73cac1e00db758beb727827c22415]]
Former-commit-id: d9a688db279c303e895e6fe7d6d7b2365e56c0b3 [formerly df9fc9f943f22f2bed0effa3d7b3fbd69cf76472]
Former-commit-id: e3611180d3d58059b6449273cda51efc07e376be

+++ Source Code/Platform/Video/OpenGL/Text/Headers/glfontstash.h
@@ -46,8 +46,8 @@ static int glfons__renderCreate(void* userPtr, int width, int height)
     gl->width = width;
     gl->height = width;
     Divide::GL_API::bindTexture(0, gl->tex, GL_TEXTURE_2D);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_R8, gl->width, gl->height, 0, GL_RED, GL_UNSIGNED_BYTE, 0);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
+    glTexImage2D(GL_TEXTURE_2D, 0, static_cast<GLint>(GL_R8), gl->width, gl->height, 0, GL_RED, GL_UNSIGNED_BYTE, 0);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, static_cast<GLint>(GL_LINEAR));
     Divide::GL_API::unbindTexture(0, GL_TEXTURE_2D);
     return 1;
 }
